Give container images labels
Appropriate labelling per http://label-schema.org/rc1/. Most of it is
static, other than the revision and build date, which are supplied when
building. Because at least the build date will be different for every
build, it appears at the end of the Dockerfiles (this used to matter;
maybe it doesn't now).