Distinguish automated release events
Previously when we ran automation the changes to be made were
collected into a set of releases of individual images. This was
largely to fit with the existing release mechanism.

The recent tag filtering implementation changed how releases could be
specified, allowing you to supply an implementation of
`update.Changes` which would then be used to calculate the changes to
manifests.

However, the commit notes and events still accommodate only the
original kind of release spec. The upshot: fields in commit notes were
empty, and notifications missing.

One quick fix would be to finesse automated release specs to look like
the original kind. However, since the former can mention multiple
images, it is easier said than done. Instead I decided to bite the
bullet and create a new kind of event (and hook it into notification).

Also removes the custom Slack template, which we didn't use.

Fixes #641.