Prefix ssh keygen tmpdir with ..weave-
The 'AtomicWriter' used by k8s to project secrets into a tmpfs volume
periodically removes any files whose names do not match existing keys in
the secret, unfortunately including the temporary directories into which
we generate keys. This can be avoided by prefixing our temporary
directories with '..', effectively placing them in the namespace
reserved by the AtomicWriter for non-user-visible files; since we are
intruding into this namespace, we also include 'weave-' in the prefix in
an attempt to avoid collisions.