Correct the serviceIDs calculation for deployments
To attribute events to particular services when syncing, we go through
all the resources in files that have changed, and ask them which
service they belong to. In the case of deployments, it's the service
that uses the pods in the deployment; but, the calculation did not
account for namespaces.