Make fluxd restarts more graceful
When flux is itself restarted (e.g., by applying a new deployment
config, perhaps from a release), the new pod will get in a kicking war
with the old pod while the latter winds down.

To mitigate this, two measures.

1. Always wait before reconnecting, rather than waiting only if there
was an error. When being kicked, the connection is shut cleanly;
previously, it would try to reconnect immediately, prompting another
kick, and so on.

2. Use the rollout strategy `Recreate` (in the example deployment
yamel). `Recreate` does what we want: stop the previous pod, start the
new one.