Record the connection time of daemons
When a daemon connects, record the time it showed up and the fact of
connection. When it disconnects, clear the latter.

A bit of care has to be taken, because when a daemon is restarted, it
can establish the connection before the old one disconnects; and, we
don't want to clear the connection flag in this case.