Parse docker credentials from imagePullSecrets
This commit moves the method that returns a list of images to warm into
the kubernetes package. This is because the method is specific to k8s
anyway and the new code to parse the secrets needed access to the k8s
cluster.

ImageCreds is a new struct to colocate the ImageID and the credentials
required to pull them.

Credentials are now added to the client via the client factory for each
new client. This is because each image might have different credentials.