Refactor kubernetes.go
The reason for this refactor is to allow testing k8s methods.
This is achieved by introducing two levels of interface. A low-level
KubeAPI, which encapsulates all the required kubernetes API calls and
a mid-level KubeFluxAdapter interface which signifies the change from
the kubernetes domain (v1.*) to the flux domain (flux.Service).

- Tests added by mocking out these interfaces
- Code moved around to make the interfaces more prominent
- Main altered to adhere to new interfaces
- AllServices and SomeServices (core api methods) have been altered
to use the new interfaces. Tests have been added.
- Sync, Export and Ping haven't been altered yet

A quick note on the SomeServices change. I decided to simply list all
services then filter for the ones requested to allow testing, simplify
the code and avoid adding more methods to the KubeAPI interface.
Previously it hit the raw k8s API to get each namespace, get a requested
service for that ns then find all `podController`s for the service. This
change results in a few more calls than necessary going to the k8s api,
but I deemed this acceptable to reduce the API and add tests.