Add flag for setting both tag and notes ref
If you are pointing more than one fluxd at a single git repo, you need
to supply a distinct `--git-sync-tag` and `--git-notes-ref`. It's easy
to forget one or other, so this commit adds a flag which, when
supplied, sets both of those values.

There will be running systems that assume that the old flags are
used (defaulted or otherwise). Therefore, only set the values if
`--git-label` is not supplied. In other words, if you are already
running fluxd, just stick with the flags you have.