Allow comments between containers when updating
This is an adjustment to the regexp that grabs containers block, which
was not accepting lines which started with `#` (i.e., comments).

I've also put a backstop in: if not every container (from the parsed
spec) that was expected to be updated was updated (in the bytes of the
definition), bail with an error.