Optimisation - reduce git note gets (#721)
* Add method to list revisions with notes

This method uses git notes list to get all notes for a given note ref.
We then turn the list into an array and take the second field from the
results (which corresponds to the object reference - the commit id in
our case). Finally, the result is placed in a map to make it easier
to do "if note is in" type queries later.

* Check if notes exist before requesting them

This is an optimisation for #714. We perform a single gitCommandExec to
get hashes for all commits with notes, place them into a map and use
them to query whether a commit has a note or not. This prevents
multiple calls to gitCommandExec just to see if there is a note
attached.

* Add extra check for error seen on ubuntu systems.

* Add context.