Push context deadlines inside jobs and sync ops
Synchronising with the cluster can easily take tens of seconds, so it
is better to have timeouts for individual operations or small batches
of operations, rather than the whole thing.

Similarly, jobs can wait in the job queue for e.g., syncs to happen,
so it is necessary to time them from when they _start_, not when they
are requested.

Lastly, fix and use `git.Repo.NoteRevList` when answering JobStatus
from commit notes, since that is exactly the expensive batch of git
operations that motivated it.