Treat non-fatal v6 errors as application errors
We don't get the full application error structs over the v6 RPC
protocol, but we do know what is a transmission error and what is a
"regular" error returned from the remote method.

We can reasonably treat the latter as application errors, which will
impart maginally better information to the user, as well as not
looking like internal server errors in logs and monitoring.