Wait for platform presence in fluxsvc test
One source of flakes appears to be the race between the platform being
registered and seeing if the platform is registered (both
`TestFluxsvc_Register` and `TestFluxsvc_Ping`).

This change forces those tests to block until the platform has
demonstrably subscribed, or fail if it does not happen within five
seconds.