Patch the notification events config to the default
If we use `omitEmpty` on the notifyEvents field, we lose the ability
to distinguish between a config that does not have a value (because
it's from before we had this config), and a config with an empty value
(`[]string{}`).

This means you cannot set the field to the empty value, since it will
be interpreted as "no such config; use the backwards-compatible
default".

Once removing the annotation, though, it causes problems for the UI,
which is not expecting `null` in its JSON response. There are three
choices at this point:

 1. change the UI to deal with `null`
 2. patch the outgoing value to be `[]` rather than `null`
 3. patch the outgoing value to be the default

Of these, 3. recommends itself since the config UI will show the
actual state in the case of an old config, which is not true of the
first two alternatives.