Remove InstanceID from api, and use contexts
We have InstanceID as the first argument to all the api interface
methods, but it's only the upstream service that cares about
instance(ID)s. Nothing else cares, and they all have to fake it.

They _do_ however care about contexts, both on the client side (for
supplying contexts to requests, even if it's `context.Background()`)
and server side (for threading contexts through from requests).

So, make the first argument of all api methods a `context.Context`
rather than an InstanceID. In the case of the service, the instance ID
is fished out of the request and supplied to the context (the _other_
use of contexts ..).

Most of the client-side contexts are TODO, until I go through and
figure out what each should be (maybe just `context.Background()`).