Use context.Background for client requests
This commit updates all the fluxctl subcommands to supply a context
when using the API. The HTTP API client uses the supplied context in
its requests.

This has not that much effect on most use of fluxctl, except that
long-running requests might be canceled should the user interrupt the
process.