Use context in remote.Platform methods
The rationale for this is not immediately obvious.

`remote.Platform` represents the RPC protocol between the service and
the daemon. As such, the daemon implements remote.Platform, and adapts
it to serve the `api.Client`.

Thus, if we want to pass contexts from API requests to the daemon,
that interface must include contexts.

The superficial benefit -- that remote calls can be given contexts --
doesn't bear out, since `net/rpc` doesn't support contexts anyway. But
if we move to another means of RPC, that benefit may accrue.