Try to get a schema2 manifest first
Version 2 of the Docker Registry API has two different schemas for
image manifests: schema1 and schema2.

Different registries deal with this differently. Google Container
Registry (gcr.io) will serve only schema2, at least for some
images. In this case we get an HTTP 404 response -- the
docker-registry-client library sets an `Accept` header withthe schema1
media type, which the registry then refuses to supply.

Thus, in this commit we ask for a schema2 image first, which is then
processed slightly differently (the data we want is in a layer, which
we have to download subsequently). If we get a 404 from that, fall
back to asking for a schema1 manifest.

It's possible to get an empty schema2 result from some registries: the
registry ignores the Accept header and sends a schema1 manifest
anyway (possibly). In this case, fall back to asking for schema1.