Parse shorthand dockerhub images correctly
This fixes issue #804 by adding test cases that were missing, and
rewriting the parsing to pass.

In addition, I have changed the use of ImageID such that it
canonicalises elements (e.g., the domain) when asked, rather than when
parsing. This means that in general, images will appear as they are
supplied by users -- e.g., `alpine:3.5`.