Use canonical image name when comparing images
The implementation of `ImageID.Repository()` was changed recently so
that it returned a canonical image _path_.

Prior to that change, it would return a minimal name for the image,
i.e., by omitting the parts implied by convention. This provided a
canonical name, in the sense of it being unambiguous provided the
conventions didn't change. It also included the registry domain e.g,
quay.io, if it was not DockerHub.

After the change, `ImageID.Repository()` gives a canonical image path
by adding in the implied parts (i.e., the prefix `library/`, for
DockerHub images) where missing, but does not ever include the
registry name.

While we were storing image metadata against the canonical
_name_ (registry and path), we were now querying using only the _path_
-- and for images not from DockerHub, nothing would be found.

The solution is to use the canonical name for storing and querying,
and the path only for the registry operations that require exactly
that.