Update gorilla/mux vendorage and adjust
To bring things in line with weaveworks/service, where these packages
are themselves vendored, update gorilla mux.

There is one hiccough: `MakeURL` no longer works as it is, since
Route.URL() now complains if it doesn't get all the parameters given
in the route, including query parameters. To work around this for the
time being, use Route.URLPath() instead, relying on the coincidence
that our routes don't contain parameters in the path.