Give the image cache warmer a priority channel
To process image push hooks, we need to be able to pre-empt the cache
warming loop.

This commit introduces a priority channel, as a field of the warmer,
and changes the loop such that it looks up image data from the
priority channel in preference to refreshing image data from the
cluster.

It is possible to be notified about an image for which the loop does
not have credentials. This means the priority lookup will be skipped,
and is otherwise harmless. The credentials are only requested at the
start, and when the regular refresh is run. A refinement which
improves on this would be to plumb through a means of getting a
specific image's credentials from the cluster.