Plumb new image notification through
When refreshing the metadata for an image (possibly at the prompting
of the priority channel), we can compare the stored set of tags with
those we have just fetched. If there are new tags in the latter,
that's a sign that it's worth doing an automation run -- the new tag
might well be an update for a controller.

Since we might miss a new image notification for whatever reason, we
still drive the automation on a ticker; this just pre-empts that
ticker.