Remove intermediary registry abstractions
It used to be the case that flux used a pull-through cache for image
metadata. The cache was practically, but not logically, necessary. It
made sense then for there to be a registry Client interface with two
implementations: the implementation that looked at the cache, and the
implementation that went and fetched the metadata from a registry.

But we changed this such that we only ever look in the cache when
querying, and refresh from registries in the background. Querying can
just use cache operations, and only refreshing needs to fetch metadata
from registries.

This commit removes some of the layers that existed only because of
the erstwhile equivalence of the two clients. Specifically: it removes
the registry struct, which used a (cache) client factory to get a
(cache) Client which it then used to assemble image
manifests. Instead, the cache client just assembles the manifests
itself.

Consequently, the cache client factory can also go. I have kept the
remote client factory around for now, as the work it does to construct
a client has to live somewhere.