Write whole results to the cache
There are two components that read or write to the cache. The first is
the cache warmer, which refreshes the list of tags for an image
repository in the cache, and for each tag, writes the image metadata
(the layer created date) to the cache; the second is `registry.Cache`
(an implementation of `registry.Registry`), which assembles a full
list of `image.Info` by first reading the tags from the cache, then
for each tag, reading the metadata.

There are two problems with this:

 - the `registry.Cache` does a lot of reads to re-assemble what the
   warmer already had all in one place; and,
 - if the warmer fails to fetch a single manifest, the Cache cannot
   assemble an answer at all.

This commit changes the warmer so that it writes whole results, and
`registry.Cache` so that it reads whole results. This solves the first
problem. The writing is done as a ratchet: we only overwrite the
result when we have managed to get all the metadata. This solves the
second problem.

A consequence is that we no longer need to store the tags separately,
as they are an intermediate result.