Rewrite registry client to use docker distribution
The package github.com/docker/distribution/registry/client (now) has
types and procedures for fetching image metadata.

This means we can get rid of a lot of the workarounds we were using to
patch `docker-registry-client` so that e.g., it works with quay.io.

I have changed the interfaces around a bit, since we usually need to
request image manifests "in a straight line", reusing the same
authorisation, and it makes sense to construct a client for each such
series of requests.

There are a few things we can keep track of across series of requests:
specifically, the challenges we've seen from each host. So it's still
useful to have a "factory" to hold that information, as well as other
commonalities like rate limiting.