Supply a reason for "available images" being blank
When flux has not managed to fetch the image metadata, it will supply
a null in the "available images" field. But usually, we can figure out
why there's no result -- it could be because we didn't have
credentials for the registry, or perhaps because we just haven't tried
to fetch them yet.

This commit effectively adds a field to the API return value to hold
that reason, and gives a placeholder reason in the case that available
images weren't found. More work is needed to plumb through the
_actual_ reasons from the registry.