Check that the upstream repo can be written to
Our "gitsync" mechanism relies on being able to push a tag back to the
upstream repository. But in the daemon, we consider it ready to go
once we've managed to clone the repository. This results in confusing
errors: if the repo is read-only (either because the deploy key is, or
because it's an `https` URL) then syncing will succeed in effect, but
report problems.

This commit adds a check that we can write to the upstream repo,
before considering the daemon ready to go. The check is that we can
push an otherwise-meaningless tag (then delete it) -- this being a
fairly innocuous write operation.