Generate helpful error on failed git write check
While starting up, the daemon continually checks whether it can clone
the upstream repo, and if so, whether it can write to it. If this
results in an error, it stores this away and keeps trying. Until it
succeeds, when asked to do anything related to the git repo (i.e.,
almost anything), it responds with the stored error.

In that interim, it's possible that a user will ask for something that
needs a cloned repo; and if so, a helpful error is .. helpful.