Modify Applier to stage and execute rather than execute directly
This change to the interface facilitates creating an Applier that combines all
yamels into a multidoc to be applied in one call to kubectl.

Note that the deleted test was testing behaviour that exists nowhere in flux.
All references to `SyncAction` in the code contain only a single change, so the
test was unnecessary.