Denormalise image refs from registry cache
When fetching image metadata, we canonicalise the image name, so that
we can fetch it using any of its possible names (e.g., alpine,
library/alpine, index.docker.io/library/alpine).

When fetching though, we must remember to reconstruct the image refs
as though they used the original form of name.

To guard against getting this wrong again, I've put the
`ImageMap` (map of image name to image metadata, effectively) in a
struct so it can't be indexed directly.