Separate out refresh loop for git mirroring
This rewrites the git package so that there are two distinct
abstractions: a self-syncing `Repo` and a working directory
`Checkout`.

`Repo` maintains a git mirror of an upstream git repository. It does
both the initial clone and regular fetches, notifying on a channel
when it's done so. The advantage of it being a mirror is that it will
keep track of all refs (including notes), so does not need to be told
to fetch particular branches.

Since the daemon often needs to push new commits, then wait for them
to be mirrored, there is also a synchronous method to fetch from
upstream.

Since it's just a mirror (i.e., has no working files), if the daemon
needs to look at files it must clone the repo. There's now a helper
method for this -- `*Daemon.WithClone` -- which makes sure things are
tidied up afterward.