Cope when auth domain differs from registry domain
It's possible to be looking up an image at index.docker.io and be told
to authorise with auth.docker.io. In this case, the registry fetch
currently fails, because it has been supplied with credentials for
index.docker.io, and is asked for credentials for auth.docker.io.

Since, at the time, we're fetching a particular image and the
credentials are associated with the image's domain, this is a failure
of logic. We should use the exact credentials collected for the
image's domain.

I've added a bit more logging such that when --registry-trace is on,
the nature and provenance of the credentials will be shown.