Symlink regenerated secret key to expected path
We now include an ssh_config file in the image, to avoid having to
supply the identity file path on the command line each time.

The way keys are generated means that it's possible the key won't
always be at the path that's in that config
`/etc/fluxd/ssh/identity`. If there's no key when the daemon starts
up, or you regenerate it, the daemon will create one in a
tmpfile. Eventually, at least on the next reboot, it will be mounted
at the expected path; but in the meantime, the ssh_config will be
looking at the old key.

To fix this, maintain the symlink at the expected path so that the
expected path is _always_ the latest key. Since we cannot update the
k8s secret and move the link atomically, this means regenerating the
key can fail partially, updating the secret but leaving the symlink
where it is. The trade-off is that the key is at a stable position.