Remove api.Client methods that aren't in remote.Platform
These changes go some of the way towards allowing us to merge the two
interfaces. Fortunately it was the case that the methods not in Platform could
easily be replaced by methods that are.

Originally this commit removed the Update{Images,Policies} and
{Get,Regenerate}PublicSSHKey routes and handlers, but a case was made for
supporting requests from older fluxctls so they were permitted to stay.