Make sure multi-yamels are delimited
We concatenate multiple manifest files into a YAML multidoc, so that
we can `kubectl apply` it all at once. The string `"---\n"` was used
to put a YAML document delimiter betwen each file's contents; however,
if a file does not end in a newline, this doens't end up on its own
line, and it interpreted as part of the preceeding document.

-> use `"\n---\n"`, on the basis that blank lines (if the file _does_
have a trailing newline) are harmless.

This commit also adds a bit more detail of what's being applied, to
the log.