Support basic auth and HTTP for image registries
The image registry code assumed
 - all image registries use HTTPS
 - all image registries use token authentication

..but if you run your own registry in your cluster, neither of these
things is likely to be true.

To support basic authentication, all we need to do is add a handler in
when constructing the registry client; it will get used if the
authentication challenge indicates so.

Supporting HTTP is (oddly) a bit trickier, since there's no indication
from an image name (which is all we have) whether the registry uses
HTTP or HTTPS. Registries will tend to redirect any HTTP
requests to HTTPS, so we _could_ try HTTP first and follow any
redirection. However, if a registry supported both HTTP and HTTPS, and
didn't redirect, we'd end up sending credentials over an insecure
connection unnecessarily.

Instead of that, make it possible to tell the daemon which registries
to use HTTP for, with the (multiply-valued) argument
`--registry-insecure-host`.