Move Upstream interface out of api.Server
Some methods on api.Server (Ping, Version, NotifyChange) only make sense in the
case that a Flux is connected to an Upstream (Weave Cloud). The result of this
was that the HTTP client ended up with 'not implemented' implementations of
methods which are out of place.

Since api.UpstreamServer is a superset of api.Server and struct embedding
exists, there luckily isn't a huge amount of change the the boilerplatey stuff.

Type assertions of the form `var _ api.Server = Foo{}` have been updated to make
sure things that should implement UpstreamServer do so.