Use kubeyaml to do image updates
This wipes out a bunch of regexp-(ab)using code, in favour of using a
self-contained binary from an image, which is expected to be on the
`$PATH`.

For the tests, we have to have a `kubectl` in the `$PATH`. A small
shell script shim suffices.

To make the binary available in a pod, the kubeyaml image can be
called as an initContainer, copying the files to a shared volume.