Avoid an NPE when parsing LogEvent responses
We assume that if the upstream we're posting events to responds with
`Content-Type: application/json`, it must be a helpful error
(`*github.com/weaveworks/flux/errors.Error`). However, it's possible
to get any old JSON (even no JSON), and in that case, we'd trip
oursevles up by creating a helpful error with a nil pointer, leading
to a panic later.