Wire kubeyaml into the fluxd deployment
To be able to use kubeyaml from within a pod, fluxd must have access
to the binaries. We can simply build the files into the image using a
multistage Dockerfile, and make sure they are on the `$PATH`.

I took the opportunity to factor out the code that executes
`kubeyaml`.