Account properly for no git config
The *git.Repo state machine has a state for `RepoNoConfig` (meaning no
repo URL supplied), but hitherto it was not reachable, because the
constructor initialised the state to `RepoNew` (not cloned yet), and
there is no transition from there to `RepoNoConfig`.

We do want to distinguish the "no config" state from just not being
ready, however. This commit makes the state reachable, and accounts
for it in *git.Repo methods.

Ultimately this is visible via the API, where "no config" is now
treated as a reason for ListController results to be read-only.