Make keygen path default to secret volume path
For Kubernetes 1.10 and onwards, people will need to supply another,
writable tmpfs volume for us to generate keys in. But we don't need to
make it a breaking change.

This commit makes the argument `--ssh-keygen-dir` optional by
defaulting to using the path of the secret volume. This means existing
configurations won't break when deploying this code. I have made the
example deployment forward-compatible -- i.e., it will work for
Kubernetes 1.10 as well as present Kubernetes releases.