Add ability to append [ci skip] to commit messages
Appending `[ci skip]` to a commit message will instruct CI to skip the
build for that commit, at least in these systems:

 - https://circleci.com/docs/1.0/skip-a-build/
 - https://docs.travis-ci.com/user/customizing-the-build/#Skipping-a-build
 - https://docs.gitlab.com/ee/ci/yaml/#skipping-jobs

So: I've added a flag which switches that will do exactly that, to all
commits fluxd creates. To account for other CI systems that have a
similar mechanism but don't use that convention, the flag
`--git-ci-skip-message` lets you supply the text to append.