Report the affected resources for each event
Hitherto we have used a single list of `ResourceID`s to populate the
event field. It was calculated by looking at which files have changed
since last sync, and which resources are in those changed files.

This is inaccurate because

 1. files don't always contain one resource (though currently, files
 _we can change_ only contain one resource)
 2. each event gets the list of resources changed in _any_ commit.

To address 2., use the Result field from the git note, which has a
record of what was actually changed in that operation.

1.) is less of a problem for now, because we can only change files
containing one resource at present anyway.