Log in places we might encounter an untagged image
We don't expect to see untagged images from the image metadata DB, but
we do. To track down #1051, this commit adds some logging to a couple
of places where we would see an empty tag, as well as taking action
when that happens (either aborting the cahe warming op, or skipping
the automated update).