Make sync run immediately once git is ready
Now that git mirroring is done in its own goroutine, something has to
trigger the sync loop to do a sync. A "cell" (channel with buffer size
1) is used to prompt the sync loop every time the repo fetches from
upstream.

However, this only happens `git-poll-interval` _after_ the mirroring
has started, and most of the time when the sync loop starts, the git
repo isn't ready. That means the first successful sync, for a
freshly-configured fluxd, is usually five minutes after starting.

To compensate, this commit makes the repo signal on the cell when it
moves to the ready state.