Resurrect the --docker-config flag for local use
We deprecated the --docker-config flag when we switched to looking up
imagePullSecrets to get credentials. However, it's still useful if
you're not using those -- for example, if you don't pull images from
the registry, or (less so) if you assign credentials to service
accounts.

This commit brings the flag back, and implements it by loading the
given docker config file, to supplement any credentials found via
imagePullSecrets.