Simplify policy and image updates internals
In the main, this rewrites Kubernetes UpdatePolicies to use the
generic resource parsing (in cluster/kubernetes/resource), rather than
doing its own. It's not possible to avoid entirely, because the way
the updating is written, it needs the annotations rather than the
policy.

As a side effect, this introduces a ResourceID into the arguments;
this will be useful for supporting multidocs and lists, later.

The internals of UpdateImage can also be rewritten, to a lesser
extent. I also removed some of the code that dealt with
ReplicationControllers, which we have not supported for a while.