Add `fluxctl sync` command
This adds the ability to manually trigger a sync (pull commits from
the git origin, and apply to cluster), and a command `fluxctl sync`
which triggers a manual sync and reports its progress.

Since a sync is potentially a long-running task, it's run using the
job machinery: the command posts an `update.Spec`, which is submitted
as a job and the ID returned. The command then proceeds more or less
as it would for a release or policy change.

Because the different kinds of update require different handling, I've
factored the creation of jobs out into some combinators (or near
combinators):

 - run a job that needs a git clone
 - run a job that results in a commit, and report the result back to
   the upstream server