Implement image updates for FluxHelmReleases
This mainly consists of making sure
cluster/kubernetes/resource.FluxHelmRelease implements
resource.Workload, and bumping the version of kubeyaml to a release
that can update FluxHelmRelease resources. That enables all the
requisite machinery, including verifying releases.

An incidental change: throwing away the (`error`) return value of
SetContainer when verifying hid a problem; so, don't ignore it.