Use const name for FHR container
When interpreting FluxHelmRelease (FHR) resources as workloads, we
have to synthesise a set of (zero or one) containers from the
information given in each FHR. The simplest way to do this is to
expect a field in the `values` for the release, called `image`, giving
an image ref to be interpolated into the chart's templates
somewhere (we don't care where).

Rather than trying to come up with a way to derive a container name
particular to the release, or chart, or chart release, this commit
just hard-codes the name to something indicating the meaning of the
value. This requires a compensating change to kubeyaml (which is
include in quay.io/squaremo/kubeyaml:0.3.1).

Additional: use type test rather than type assertion, so it will
return `nil` rather than exploding.