Use ~/.kube/config to set default namespace
`kubectl` run from the pod uses the service account token, and
inherits its default namespace from the service account. We want it to
be the default default (named "default"), to mimic the behaviour of
`kubectl` used from a terminal and thereby avoid some surprises.

To reset the default namespace, we copy a Kubernetes config file with
a context specifying the "default" namespace, to the place `kubectl`
expects to find it, in the flux image.