Write whole file updates at a time
Up to this point, the update (and release) packages assumed that each
file they were dealing with corresponded to a single manifest, because
we couldn't deal with updating files with multiple manifests in them.

Since we can now update a resource in a multidoc file or List, we can
do this correctly and more simply by just overwriting each file after
running its contents through the update process, serially.