Put baked-in SSH config in /etc/ssh
At present we bake SSH config (ssh_config and known_hosts) into the
flux image at the user-specific location, that is `/root/.ssh/`. This
makes it fiddly to override by mounting a configmap, because you have
to specify subPath (which stops Kubernetes from updating the file when
the configmap changes), or supply both files.

If we put our baked-in config at the global location (`/etc/ssh`),
people are free to mount a configmap at `/root/.ssh` without the
limitations above. This means we can unconditionally mount a (probably
empty) configmap in the example and in Helm charts.