Clean up fully after git clones
For both mirrors (in git/repo.go) and working clones (in
git/working.go) we created a temp directory, then cloned into a
_subdirectory_ of that. But when tidying up, we deleted the
subdirectory -- leaving lots of empty temp directories around.

This changes mirroring and cloning so that they simply use the templ
directory. It also adds a `Clean` method to Repo, so that they can be
cleaned up after tests. Result: you should see no /tmp/flux-git*
directories after running the tests.

(This has the same changes as cf85654, because we want to reapply it
after a CI hiccough.)